Move ValidateContextDirectory to the one package that uses it.
