+++ api/client/utils_windows.go
@@ -1,6 +1,6 @@
 // +build windows
 
-package utils
+package client
 
 import (
 	"path/filepath"
